requirements.txt
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2

openai-whisper

sacremoses

TTS==0.22.0

soundfile
librosa
scipy

tqdm
regex
requests

transformers==4.40.2
tokenizers==0.19.1
accelerate==0.29.3
sentencepiece
protobuf